+++ server/api/index.js
@@ -2,7 +2,7 @@ const router = require('express').Router()
 module.exports = router
 
 router.use('/products', require('./products'))
-
+router.use('/cart', require('./cart'))
 router.use('/users', require('./users'))
 
 router.use((req, res, next) => {
